--- a/src/Microsoft.OData.Core/Json/BufferingJsonReader.cs
+++ b/src/Microsoft.OData.Core/Json/BufferingJsonReader.cs
@@ -248,7 +248,7 @@ namespace Microsoft.OData.Json
         /// <returns>A task that represents the asynchronous read operation.
         /// The value of the TResult parameter contains the node value of the last buffered read if buffering is on or off;
         /// otherwise the node value of the last unbuffered read.</returns>
-        public Task<object> GetValueAsync()
+        public ValueTask<object> GetValueAsync()
         {
             this.AssertAsynchronous();
 
@@ -257,11 +257,11 @@ namespace Microsoft.OData.Json
                 if (this.isBuffering)
                 {
                     Debug.Assert(this.currentBufferedNode != null, "this.currentBufferedNode != null");
-                    return Task.FromResult(this.currentBufferedNode.Value);
+                    return ValueTask.FromResult(this.currentBufferedNode.Value);
                 }
 
                 // In non-buffering mode if we have buffered nodes satisfy the request from the first node there
-                return Task.FromResult(this.bufferedNodesHead.Value);
+                return ValueTask.FromResult(this.bufferedNodesHead.Value);
             }
 
             return this.innerReader.GetValueAsync();
@@ -272,7 +272,7 @@ namespace Microsoft.OData.Json
         /// </summary>
         /// <returns>A task that represents the asynchronous read operation.
         /// The value of the TResult parameter contains true if a new node was found, or false if end of input was reached.</returns>
-        public Task<bool> ReadAsync()
+        public ValueTask<bool> ReadAsync()
         {
             Debug.Assert(!this.parsingInStreamError, "!this.parsingInStreamError");
             this.AssertAsynchronous();
@@ -286,7 +286,7 @@ namespace Microsoft.OData.Json
         /// </summary>
         /// <returns>A task that represents the asynchronous operation.
         /// true if the current value can be streamed; otherwise false.</returns>
-        public virtual async Task<bool> CanStreamAsync()
+        public virtual async ValueTask<bool> CanStreamAsync()
         {
             this.AssertAsynchronous();
 
@@ -307,7 +307,7 @@ namespace Microsoft.OData.Json
         /// <returns>A task that represents the asynchronous read operation.
         /// The value of the TResult parameter contains a <see cref="Stream"/> used to read a stream value.</returns>
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1825:Avoid zero-length array allocations.", Justification = "<Pending>")]
-        public virtual async Task<Stream> CreateReadStreamAsync()
+        public virtual async ValueTask<Stream> CreateReadStreamAsync()
         {
             this.AssertAsynchronous();
 
@@ -342,7 +342,7 @@ namespace Microsoft.OData.Json
         /// </summary>
         /// <returns>A task that represents the asynchronous read operation.
         /// The value of the TResult parameter contains a <see cref="TextReader"/> for reading the text value.</returns>
-        public virtual async Task<TextReader> CreateTextReaderAsync()
+        public virtual async ValueTask<TextReader> CreateTextReaderAsync()
         {
             this.AssertAsynchronous();
 
@@ -682,7 +682,7 @@ namespace Microsoft.OData.Json
         /// represents an in-stream error. If so, it throws an <see cref="ODataErrorException"/>. If false, this check will not happen.
         /// This parsingInStremError field is set to true when trying to parse an in-stream error; in normal operation it is false.
         /// </remarks>
-        protected async Task<bool> ReadInternalAsync()
+        protected async ValueTask<bool> ReadInternalAsync()
         {
             this.AssertAsynchronous();