--- a/src/Microsoft.OData.Core/Json/IJsonReader.cs
+++ b/src/Microsoft.OData.Core/Json/IJsonReader.cs
@@ -68,7 +68,7 @@ namespace Microsoft.OData.Json
         /// </summary>
         /// <returns>A task that represents the asynchronous operation.
         /// The value of the TResult parameter contains true if a new node was found, or false if end of input was reached.</returns>
-        Task<bool> ReadAsync();
+        ValueTask<bool> ReadAsync();
 
         /// <summary>
         /// Asynchronously returns the value of the last reported node.
@@ -85,27 +85,27 @@ namespace Microsoft.OData.Json
         /// - Double if a number which doesn't fit into Int32 was found.
         /// If the last node is a Property this property returns a string which is the name of the property.
         /// </remarks>
-        Task<object> GetValueAsync();
+        ValueTask<object> GetValueAsync();
 
         /// <summary>
         /// Asynchronously creates a Stream for reading a binary value.
         /// </summary>
         /// <returns>A task that represents the asynchronous operation.
         /// The value of the TResult parameter contains a <see cref="Stream"/> for reading the binary value.</returns>
-        Task<Stream> CreateReadStreamAsync();
+        ValueTask<Stream> CreateReadStreamAsync();
 
         /// <summary>
         /// Asynchronously creates a TextReader for reading a string value.
         /// </summary>
         /// <returns>A task that represents the asynchronous operation.
         /// The value of the TResult parameter contains a <see cref="TextReader"/> for reading the string value.</returns>
-        Task<TextReader> CreateTextReaderAsync();
+        ValueTask<TextReader> CreateTextReaderAsync();
 
         /// <summary>
         /// Asynchronously checks whether or not the current value can be streamed.
         /// </summary>
         /// <returns>A task that represents the asynchronous operation.
         /// true if the current value can be streamed; otherwise false.</returns>
-        Task<bool> CanStreamAsync();
+        ValueTask<bool> CanStreamAsync();
     }
 }