… src/Microsoft.OData.Core/Json/JsonReader.cs
… src/Microsoft.OData.Core/Json/ODataJsonWriterUtils.cs
… src/Microsoft.OData.Core/Json/ODataUtf8JsonWriter.Stream.cs
… src/Microsoft.OData.Core/Json/TranscodingWriteStream.cs
… src/Microsoft.OData.Core/JsonLight/JsonLightODataAnnotationWriter.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightBatchWriter.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightCollectionWriter.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightDeltaWriter.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightInputContext.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightOutputContext.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightPropertySerializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightReader.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightResourceDeserializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightSerializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightValueSerializer.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightWriter.cs
… src/Microsoft.OData.Core/JsonLight/ODataJsonLightWriterUtils.cs
… src/Microsoft.OData.Core/JsonLight/ReorderingJsonReader.cs
… src/Microsoft.OData.Core/Microsoft.OData.Core.csproj
… src/Microsoft.OData.Core/MultipartMixed/ODataMultipartMixedBatchInputContext.cs
… src/Microsoft.OData.Core/MultipartMixed/ODataMultipartMixedBatchOutputContext.cs
… src/Microsoft.OData.Core/ODataDeltaWriter.cs
… src/Microsoft.OData.Core/ODataInputContext.cs
… src/Microsoft.OData.Core/ODataMessageReader.cs
… src/Microsoft.OData.Core/ODataMessageWriter.cs
… src/Microsoft.OData.Core/ODataMessageWriterSettings.cs
… src/Microsoft.OData.Core/ODataMetadataJsonOutputContext.cs
… src/Microsoft.OData.Core/ODataMetadataOutputContext.cs
… src/Microsoft.OData.Core/ODataOutputContext.cs
… src/Microsoft.OData.Core/ODataRawInputContext.cs