@habbes 2024-05-22
Closing because I've picked up the work in a different PR: https://github.com/OData/odata.net/pull/2975